Update to Io 20060830
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Languages@925 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Io/IoVM/base/List_inline.h
@@ -113,7 +113,7 @@ IOINLINE ListCursor List_cursor(List *self)
 
 IOINLINE void *ListCursor_next(ListCursor *self)
 {
-	if (self->index == List_size(self->list))
+	if (self->index == (size_t)List_size(self->list))
 	{
 		return 0x0;
 	}
